Updated Navigation - Added the Login & SignUp screens to the AuthNav - Created Navigation.tsx as parent nav - Updated App with the new logic implemented inside Navigation.tsx
